+++ app/src/main/java/com/projects/moviemanager/common/ui/components/bottomsheet/SortButtonBottomSheet.kt
@@ -13,7 +13,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 import com.projects.moviemanager.R
-import com.projects.moviemanager.common.ui.util.UiConstants.SMALL_MARGIN
+import com.projects.moviemanager.common.util.UiConstants.SMALL_MARGIN
 
 @Composable
 fun SortButton(
