+++ app/src/main/java/com/projects/moviemanager/common/ui/components/card/ImageContentCard.kt
@@ -14,10 +14,10 @@ import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import com.projects.moviemanager.common.domain.models.util.MediaType
 import com.projects.moviemanager.common.ui.components.NetworkImage
-import com.projects.moviemanager.common.ui.util.UiConstants.DEFAULT_PADDING
-import com.projects.moviemanager.common.ui.util.UiConstants.POSTER_ASPECT_RATIO_MULTIPLY
+import com.projects.moviemanager.common.util.UiConstants.DEFAULT_PADDING
+import com.projects.moviemanager.common.util.UiConstants.POSTER_ASPECT_RATIO_MULTIPLY
 import com.projects.moviemanager.common.domain.models.content.GenericContent
-import com.projects.moviemanager.util.Constants.BASE_300_IMAGE_URL
+import com.projects.moviemanager.common.util.Constants.BASE_300_IMAGE_URL
 
 @Composable
 fun ImageContentCard(
