+++ app/src/main/java/com/projects/moviemanager/features/browse/ui/BrowseView.kt
@@ -39,17 +39,17 @@ import com.projects.moviemanager.common.ui.MainViewModel
 import com.projects.moviemanager.common.ui.components.ComponentPlaceholder
 import com.projects.moviemanager.common.ui.components.card.DefaultContentCard
 import com.projects.moviemanager.common.ui.theme.RoundCornerShapes
-import com.projects.moviemanager.common.ui.util.UiConstants.BROWSE_CARD_PADDING_HORIZONTAL
-import com.projects.moviemanager.common.ui.util.UiConstants.BROWSE_CARD_PADDING_VERTICAL
-import com.projects.moviemanager.common.ui.util.UiConstants.BROWSE_MIN_CARD_WIDTH
-import com.projects.moviemanager.common.ui.util.UiConstants.BROWSE_SCAFFOLD_HEIGHT_OFFSET
-import com.projects.moviemanager.common.ui.util.UiConstants.DEFAULT_MARGIN
-import com.projects.moviemanager.common.ui.util.UiConstants.POSTER_ASPECT_RATIO_MULTIPLY
-import com.projects.moviemanager.common.ui.util.UiConstants.SMALL_MARGIN
-import com.projects.moviemanager.common.ui.util.UiConstants.SMALL_PADDING
-import com.projects.moviemanager.common.ui.util.calculateCardsPerRow
-import com.projects.moviemanager.common.ui.util.dpToPx
-import com.projects.moviemanager.common.ui.util.pxToDp
+import com.projects.moviemanager.common.util.UiConstants.BROWSE_CARD_PADDING_HORIZONTAL
+import com.projects.moviemanager.common.util.UiConstants.BROWSE_CARD_PADDING_VERTICAL
+import com.projects.moviemanager.common.util.UiConstants.BROWSE_MIN_CARD_WIDTH
+import com.projects.moviemanager.common.util.UiConstants.BROWSE_SCAFFOLD_HEIGHT_OFFSET
+import com.projects.moviemanager.common.util.UiConstants.DEFAULT_MARGIN
+import com.projects.moviemanager.common.util.UiConstants.POSTER_ASPECT_RATIO_MULTIPLY
+import com.projects.moviemanager.common.util.UiConstants.SMALL_MARGIN
+import com.projects.moviemanager.common.util.UiConstants.SMALL_PADDING
+import com.projects.moviemanager.common.util.calculateCardsPerRow
+import com.projects.moviemanager.common.util.dpToPx
+import com.projects.moviemanager.common.util.pxToDp
 import com.projects.moviemanager.common.domain.models.content.DetailedMediaInfo
 import com.projects.moviemanager.common.domain.models.content.MovieDetailsInfo
 import com.projects.moviemanager.common.domain.models.content.ShowDetailsInfo
