+++ app/src/main/java/com/projects/moviemanager/features/details/ui/components/moreoptions/MoreOptionsTab.kt
@@ -6,7 +6,7 @@ import com.projects.moviemanager.common.domain.models.util.MediaType
 import com.projects.moviemanager.common.ui.components.GridContentList
 import com.projects.moviemanager.common.ui.components.tab.GenericTabRow
 import com.projects.moviemanager.common.ui.components.tab.setupTabs
-import com.projects.moviemanager.common.ui.util.UiConstants.MAX_COUNT_DETAILS_CARDS
+import com.projects.moviemanager.common.util.UiConstants.MAX_COUNT_DETAILS_CARDS
 import com.projects.moviemanager.common.domain.models.content.MediaContent
 import com.projects.moviemanager.common.domain.models.content.Videos
 import com.projects.moviemanager.features.details.ui.components.moreoptions.MoreOptionsTabItem.MoreLikeThisTab
