+++ app/src/main/java/com/projects/moviemanager/features/search/domain/SearchInteractor.kt
@@ -7,7 +7,7 @@ import com.projects.moviemanager.common.domain.models.util.MediaType
 import com.projects.moviemanager.common.domain.models.content.GenericContent
 import com.projects.moviemanager.features.search.ui.paging.SearchPagingSource
 import com.projects.moviemanager.network.repository.search.SearchRepository
-import com.projects.moviemanager.util.Constants
+import com.projects.moviemanager.common.util.Constants
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
