+++ app/src/main/java/com/projects/moviemanager/features/search/ui/SearchViewModel.kt
@@ -8,7 +8,7 @@ import androidx.paging.PagingData
 import androidx.paging.cachedIn
 import com.projects.moviemanager.common.domain.models.content.GenericContent
 import com.projects.moviemanager.common.domain.models.util.MediaType
-import com.projects.moviemanager.common.ui.util.UiConstants.SEARCH_DEBOUNCE_TIME_MS
+import com.projects.moviemanager.common.util.UiConstants.SEARCH_DEBOUNCE_TIME_MS
 import com.projects.moviemanager.features.search.domain.SearchInteractor
 import com.projects.moviemanager.features.search.events.SearchEvent
 import com.projects.moviemanager.features.search.ui.components.SearchTypeFilterItem
