+++ app/src/main/java/com/projects/moviemanager/features/watchlist/ui/components/WatchlistBottomSheet.kt
@@ -13,7 +13,7 @@ import com.projects.moviemanager.common.domain.models.util.MediaType
 import com.projects.moviemanager.common.ui.MainViewModel
 import com.projects.moviemanager.common.ui.components.bottomsheet.GenericBottomSheet
 import com.projects.moviemanager.common.ui.components.bottomsheet.SortButton
-import com.projects.moviemanager.util.Constants.UNSELECTED_OPTION_INDEX
+import com.projects.moviemanager.common.util.Constants.UNSELECTED_OPTION_INDEX
 
 @Composable
 fun WatchlistSortBottomSheet(
