+++ app/src/main/java/com/projects/moviemanager/network/services/search/SearchService.kt
@@ -5,8 +5,8 @@ import com.projects.moviemanager.network.models.content.common.MultiResponse
 import com.projects.moviemanager.network.models.content.common.PersonResponse
 import com.projects.moviemanager.network.models.content.common.ShowResponse
 import com.projects.moviemanager.network.models.search.SearchPageResponse
-import com.projects.moviemanager.util.Constants
-import com.projects.moviemanager.util.Constants.ENGLISH_LANGUAGE_CODE
+import com.projects.moviemanager.common.util.Constants
+import com.projects.moviemanager.common.util.Constants.ENGLISH_LANGUAGE_CODE
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Query
